sdk: refactoring the client secret authorizer out
