+++ sdk/auth/auth_client_secret.go
@@ -2,6 +2,10 @@ package auth
 
 import (
 	"context"
+	"fmt"
+	"golang.org/x/oauth2"
+	"net/url"
+	"strings"
 
 	"github.com/hashicorp/go-azure-sdk/sdk/environments"
 )
@@ -31,5 +35,76 @@ func NewClientSecretAuthorizer(ctx context.Context, options ClientSecretAuthoriz
 		TokenVersion:       options.TokenVersion,
 	}
 
-	return conf.TokenSource(ctx, ClientCredentialsSecretType), nil
+	return conf.TokenSource(ctx, clientCredentialsSecretType), nil
+}
+
+type clientSecretAuthorizer struct {
+	ctx  context.Context
+	conf *clientCredentialsConfig
+}
+
+func (a *clientSecretAuthorizer) Token() (*oauth2.Token, error) {
+	if a.conf == nil {
+		return nil, fmt.Errorf("could not request token: conf is nil")
+	}
+
+	v := url.Values{
+		"client_id":     {a.conf.ClientID},
+		"client_secret": {a.conf.ClientSecret},
+		"grant_type":    {"client_credentials"},
+	}
+
+	if a.conf.TokenVersion == TokenVersion1 {
+		v["resource"] = []string{a.conf.ResourceUrl}
+	} else {
+		v["scope"] = []string{strings.Join(a.conf.Scopes, " ")}
+	}
+
+	tokenUrl := a.conf.TokenURL
+	if tokenUrl == "" {
+		tokenUrl = tokenEndpoint(a.conf.Environment.AzureADEndpoint, a.conf.TenantID, a.conf.TokenVersion)
+	}
+
+	return clientCredentialsToken(a.ctx, tokenUrl, &v)
+}
+
+// AuxiliaryTokens returns additional tokens for auxiliary tenant IDs, for use in multi-tenant scenarios
+func (a *clientSecretAuthorizer) AuxiliaryTokens() ([]*oauth2.Token, error) {
+	if a.conf == nil {
+		return nil, fmt.Errorf("could not request token: conf is nil")
+	}
+
+	tokens := make([]*oauth2.Token, 0)
+
+	if len(a.conf.AuxiliaryTenantIDs) == 0 {
+		return tokens, nil
+	}
+
+	for _, tenantId := range a.conf.AuxiliaryTenantIDs {
+		v := url.Values{
+			"client_id":     {a.conf.ClientID},
+			"client_secret": {a.conf.ClientSecret},
+			"grant_type":    {"client_credentials"},
+		}
+
+		if a.conf.TokenVersion == TokenVersion1 {
+			v["resource"] = []string{a.conf.ResourceUrl}
+		} else {
+			v["scope"] = []string{strings.Join(a.conf.Scopes, " ")}
+		}
+
+		tokenUrl := a.conf.TokenURL
+		if tokenUrl == "" {
+			tokenUrl = tokenEndpoint(a.conf.Environment.AzureADEndpoint, tenantId, a.conf.TokenVersion)
+		}
+
+		token, err := clientCredentialsToken(a.ctx, tokenUrl, &v)
+		if err != nil {
+			return tokens, err
+		}
+
+		tokens = append(tokens, token)
+	}
+
+	return tokens, nil
 }
